기본 게시판 완료 및, 게시판 -> 상품 게시판 변경중

+++ src/main/java/com/example/hhmm/Post/Post.java
@@ -49,7 +49,7 @@ public class Post {
     private int viewCount;
 
     @Column(nullable = false)
-    private int starpoint;
+    private float starpoint;
 
     @Column(nullable = true)
     private String imgUrl;
