+++ src/main/java/com/example/hhmm/Post/PostDTO.java
@@ -30,7 +30,7 @@ public class PostDTO {
     private LocalDateTime regDate;
     private LocalDateTime updateDate;
     private int viewCount;
-    private int starpoint;
+    private float starpoint;
     private List<CommentDTO> commentDTOs;
 
     public PostDTO(Post post) {
